+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 * Removed deprecated ``formdata``
 * Renamed ``Email Notification Form`` to ``Form (Advanced)``
+* Optimized admin export templates
 
 2.0.4 (2016-01-20)
 ------------------