+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 * Optimized admin export templates
 * Add stripped default django templates to ``/aldryn_forms/templates``
 * Implement "Advanced Settings" when configuring plugins
+* Adapt default setting ``show_all_recipients`` for aldryn users
 
 2.0.4 (2016-01-20)
 ------------------