+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 * Add stripped default django templates to ``/aldryn_forms/templates``
 * Implement "Advanced Settings" when configuring plugins
 * Adapt default setting ``show_all_recipients`` for aldryn users
+* Removed not required options from form fields
 
 2.0.4 (2016-01-20)
 ------------------