+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Implement "Advanced Settings" when configuring plugins
 * Adapt default setting ``show_all_recipients`` for aldryn users
 * Removed not required options from form fields
+* Set default for "Field is required" to ``False``
 
 2.0.4 (2016-01-20)
 ------------------