+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Adapt default setting ``show_all_recipients`` for aldryn users
 * Removed not required options from form fields
 * Set default for "Field is required" to ``False``
+* Fix Django 1.9 issues
 
 2.0.4 (2016-01-20)
 ------------------